re-organise tests
