--- packages/cookie-banner/__tests__/playwright/playwright.spec.js
@@ -1,19 +1,0 @@
-const { test, expect } = require('@playwright/test');
-import AxeBuilder from '@axe-core/playwright';
-
-let tabKey;
-
-test.beforeEach(async ({ page }, testInfo) => {
-	await page.goto('/');
-	tabKey = testInfo.project.use.defaultBrowserType === 'webkit'
-			? "Alt+Tab"
-			: "Tab";
-});
-
-test.describe('Cookie banner > Axe', { tag: '@reduced'}, () => {
-	test('Should not have any automatically detectable accessibility issues', async ({ page }) => {	
-		const accessibilityScanResults = await new AxeBuilder({ page }).analyze(); 
-		expect(accessibilityScanResults.violations).toEqual([]);
-	});
-});
-
